+++ src/main/java/com/cestec/cestec/model/funcionarioDTO.java
@@ -1,5 +1,6 @@
 package com.cestec.cestec.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.AllArgsConstructor;
@@ -12,18 +13,18 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class funcionarioDTO {
-    private Integer   codfunc;
-    private String    nomefunc;
-    private String    cpf;
-    private String    numtel;
-    private Integer   codsetor;
-    private String    nomesetor;
-    private Integer   codcargo;
-    private String    nomecargo;
-    private String    endereco;
-    private double    salario;
-    private LocalDate datinasc;
-    private LocalDate criado_em;
+    private Integer    codfunc;
+    private String     nomefunc;
+    private String     cpf;
+    private String     numtel;
+    private Integer    codsetor;
+    private String     nomesetor;
+    private Integer    codcargo;
+    private String     nomecargo;
+    private String     endereco;
+    private BigDecimal salario;
+    private LocalDate  datinasc;
+    private LocalDate  criado_em;
 
     public funcionarioDTO(Integer codfunc, String nomefunc, String cpf, String numtel, Integer codsetor, String nomesetor, Integer codcargo, String nomecargo) {
         this.codfunc   = codfunc;
